Move logging method back to module.json
