+++ scripts/chat-card.js
@@ -1,4 +1,5 @@
-import { log, spellLevelToOrder } from "./utils.js";
+import { spellLevelToOrder } from "./utils.js";
+import { log } from "./module.js";
 
 export const preDisplayPowerCard = function(item, chatData) {
     if (item.system.preparation.mode !== 'talent') {
